@@ -1,22 +1,18 @@
-import { ProfileComponent } from './Components/profile/profile.component';
+import { ProfileComponent } from './Components/Pages/profile/profile.component';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomepageComponent } from './Components/homepage/homepage.component';
-import { LoginComponent } from './Components/login/login.component';
-import { NotfoundComponent } from './Components/notfound/notfound.component';
-import { RegisterComponent } from './Components/register/register.component';
-import { SummaryComponent } from './Components/summary/summary.component';
+import { HomepageComponent } from './Components/Pages/homepage/homepage.component';
+import { LoginComponent } from './Components/Pages/login/login.component';
+import { NotfoundComponent } from './Components/Pages/notfound/notfound.component';
+import { RegisterComponent } from './Components/Pages/register/register.component';
+import { SummaryComponent } from './Components/Pages/summary/summary.component';
+import { TransferComponent } from './Components/Pages/transfer/transfer.component';
 
 const routes: Routes = [
   
   {
     path: 'login',
     component: LoginComponent
-  },
-
-  {
-    path: '',
-    component: HomepageComponent
   },
 
   {
@@ -31,6 +27,22 @@ const routes: Routes = [
   {
     path:'profile',
     component: ProfileComponent
+  },
+
+  {
+    path:'homepage',
+    component: HomepageComponent
+  },
+  
+  { 
+    path: '',   
+    redirectTo: '/homepage', 
+    pathMatch: 'full' 
+  },
+  
+  {
+    path: 'transfer',
+    component: TransferComponent
   },
 
   {
